requirements.txt
# Backend chính
flask
flask-cors
pandas
scikit-learn
gunicorn
requests

# Violence detector
torch
transformers
Pillow
opencv-python
